requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
bcrypt==5.0.0
cffi==2.0.0
click==8.3.1
cryptography==46.0.5
ecdsa==0.19.1
fastapi==0.133.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
passlib==1.7.4
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.47
sqlmodel==0.0.37
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
